% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/itm.R
\name{itm}
\alias{itm}
\title{Information Consistency-Based Measures for Spatial Association}
\usage{
itm(
  formula,
  data,
  method = c("vm", "icm"),
  beta = 1,
  unit = c("e", "2", "10"),
  seed = 42,
  permutation_number = 999
)
}
\arguments{
\item{formula}{A formula.}

\item{data}{A \code{data.frame}, \code{tibble} or \code{sf} object of observation data.}

\item{method}{(optional) whether \code{vm}(default) or \code{icm}.}

\item{beta}{(optional) The $\beta$ value used fo \code{vm} measure, default is \code{1}.}

\item{unit}{(optional) Logarithm base, default is \code{e}.}

\item{seed}{(optional) Random number seed, default is \code{42}.}

\item{permutation_number}{(optional) Number of Random Permutations, default is \code{999}.}
}
\value{
A \code{tibble}.
}
\description{
Information Consistency-Based Measures for Spatial Association
}
\examples{
d = c(rep(1,7),rep(2,3),rep(1,3),rep(2,7))
s = rep(1:2, each = 10)
demodf = data.frame(d = d, s = s)
\donttest{
itm(d ~ s, data = demodf, method = 'icm')
}
}
